feat(price-calculation-service): add DTOs, define API
