+++ price-calculation-service/pom.xml
@@ -8,7 +8,7 @@
         <version>3.4.0</version>
         <relativePath/>
     </parent>
-    <groupId>com.example</groupId>
+    <groupId>com.cabaggregator</groupId>
     <artifactId>price-calculation-service</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <name>price-calculation-service</name>
@@ -17,12 +17,17 @@
     <properties>
         <java.version>21</java.version>
         <postgresql.version>42.7.3</postgresql.version>
+        <mapstruct.version>1.5.5.Final</mapstruct.version>
     </properties>
 
     <dependencies>
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-web</artifactId>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
@@ -40,6 +45,18 @@
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.mapstruct</groupId>
+            <artifactId>mapstruct</artifactId>
+            <version>${mapstruct.version}</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.mapstruct</groupId>
+            <artifactId>mapstruct-processor</artifactId>
+            <version>${mapstruct.version}</version>
             <scope>provided</scope>
         </dependency>
         <dependency>
