+++ price-calculation-service/src/main/java/com/cabaggregator/pricecalculationservice/PriceCalculationServiceApplication.java
@@ -1,4 +1,4 @@
-package com.example.pricecalculationservice;
+package com.cabaggregator.pricecalculationservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
