+++ price-calculation-service/src/main/java/com/cabaggregator/pricecalculationservice/core/dto/PriceCalculationRequest.java
@@ -1,0 +1,8 @@
+package com.cabaggregator.pricecalculationservice.core.dto;
+
+public record PriceCalculationRequest(
+        Long distance,
+        Long duration,
+        String fare
+) {
+}
