+++ price-calculation-service/src/main/java/com/cabaggregator/pricecalculationservice/repository/WeatherCoefficientRepository.java
@@ -1,6 +1,6 @@
-package com.example.pricecalculationservice.repository;
+package com.cabaggregator.pricecalculationservice.repository;
 
-import com.example.pricecalculationservice.entity.WeatherCoefficient;
+import com.cabaggregator.pricecalculationservice.entity.WeatherCoefficient;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
