+++ price-calculation-service/src/test/java/com/cabaggregator/pricecalculationservice/PriceCalculationServiceApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.pricecalculationservice;
+package com.cabaggregator.pricecalculationservice;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
